+++ src/Display/Column/Editable/EditableColumn.php
@@ -7,6 +7,7 @@ use SleepingOwl\Admin\Display\Column\NamedColumn;
 
 class EditableColumn extends NamedColumn
 {
+
     /**
      * @var bool
      */
@@ -38,9 +39,9 @@ class EditableColumn extends NamedColumn
      * @param             $name
      * @param             $label
      */
-    public function __construct($name, $label = null)
+    public function __construct($name, $label = null, $small = null)
     {
-        parent::__construct($name, $label);
+        parent::__construct($name, $label, $small);
 
         $this->clearHtmlAttributes();
     }
@@ -153,15 +154,17 @@ class EditableColumn extends NamedColumn
      */
     public function isReadonly()
     {
+
+        // Add policy
         if ($this->getModelConfiguration()->isEditable($this->getModel())) {
             if (is_callable($this->readonlyEditable)) {
                 return (bool) call_user_func($this->readonlyEditable, $this->getModel());
             }
-        } else {
-            return false;
+
+            return (bool) $this->readonlyEditable;
         }
 
-        return (bool) $this->readonlyEditable;
+        return true;
     }
 
     /**
@@ -181,10 +184,11 @@ class EditableColumn extends NamedColumn
      */
     public function toArray()
     {
+
         return parent::toArray() + [
             'id' => $this->getModel()->getKey(),
             'value' => $this->getModelValue(),
-            'isEditable' => $this->isReadonly(),
+            'isReadonly' => $this->isReadonly(),
             'url' => $this->getUrl(),
             'title' => $this->getTitle(),
             'mode' => $this->getEditableMode(),